# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pawan Coaching Management API",
    "version": "0.1.0",
    "description": "Cookie-authenticated REST contract for the coaching management prototype."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment"
    }
  ],
...
  "paths": {
...
    "/api/students": {
      "get": {
        "tags": [
          "People"
        ],
        "summary": "List students",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor page of students"
          }
        }
      },
      "post": {
        "tags": [
          "People"
        ],
        "summary": "Create a student account and profile",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "responses": {
          "201": {
            "description": "Student created"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/tutors": {
      "get": {
        "tags": [
          "People"
        ],
        "summary": "List tutors",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor page of tutors"
          }
        }
      },
      "post": {
        "tags": [
          "People"
        ],
        "summary": "Create a tutor account and profile",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "responses": {
          "201": {
            "description": "Tutor created"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/fee-plans": {
      "get": {
        "tags": [
          "Fees"
        ],
        "summary": "Read fee-plan history",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Fee plans"
          }
        }
      },
      "post": {
        "tags": [
          "Fees"
        ],
        "summary": "Create an effective-dated fee-plan version",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "responses": {
          "201": {
            "description": "Fee plan created"
          }
        }
      }
    },
...
    "/api/fee-dues/{id}/payment": {
      "post": {
        "tags": [
          "Fees"
        ],
        "summary": "Record or update a full payment and receipt",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment and receipt snapshot"
          }
        }
      }
    },
...
    "/api/batches/{id}/enrollments": {
      "post": {
        "tags": [
          "Batches"
        ],
        "summary": "Create a dated batch enrollment",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Enrollment created with capacity and timetable warnings"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "psc_session"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token"
      }
    },
    "responses": {
      "Problem": {
        "description": "RFC 9457 Problem Details",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  }
}